remove need for 2to3 in chirp
