+++ chirp/src/bindings/python3/Makefile
@@ -28,12 +28,8 @@ chirp_wrap.c CChirp.py: chirp.i chirp.binding.py
 
 $(CHIRPPYTHONSO): $(OBJECTS) $(EXTERNAL_DEPENDENCIES)
 
-chirp.binding.py: %.py: ../python2/%.py
-	    cp $< $@
-		    $(CCTOOLS_PYTHON3_2TO3) --nobackups --no-diffs --write $@
-
 clean:
-	rm -f $(OBJECTS) $(TARGETS) Python.framework chirp.py chirp.binding.py CChirp.py chirp_wrap.c *.pyc
+	rm -f $(OBJECTS) $(TARGETS) Python.framework chirp.py CChirp.py chirp_wrap.c *.pyc
 
 test:
